refactor folders for pypi package
