+++ .github/workflows/python-package.yml
@@ -30,7 +30,7 @@ jobs:
           poetry install
       - name: Check flake8
         run: |
-          poetry run flake8 fastorm --count --show-source --statistics
+          poetry run flake8 fast_orm --count --show-source --statistics
       - name: Check isort
         run: |
           poetry run isort ./ --check-only --verbose
