+++ fast_orm/common/enums.py
