+++ fast_orm/common/utils.py
