+++ fast_orm/core/codegen/model.py
