+++ pyproject.toml
@@ -1,5 +1,5 @@
 [tool.poetry]
-name = "fastorm"
+name = "fast-orm"
 version = "0.1.0"
 description = "EdgeDB inspired ORM for PostgreSQL"
 authors = ["Ilya Lure <L1mple@rambler.ru>"]
@@ -7,7 +7,7 @@ license = "MIT"
 readme = "README.md"
 
 [tool.poetry.scripts]
-fastorm = "fastorm.cli.cli:app"
+fastorm = "fast_orm.cli.cli:app"
 
 [tool.poetry.dependencies]
 python = "^3.11"
